requirements.txt
easydict
PyYAML
pytest
scipy
numpy
pathlib
pydicom
matplotlib
ipywidgets==7.5.0
torch
torchio